tests: gdppay: refactor and add missing bits
Most of the tests weren't updated after the sticky events order
and stream start. Fix that and refactor those tests check that
are the same to some common functions.

Those functions still don't actually test the content but at
least now they are in a single place and can be improved
without replication